deleting offsprings types until further notice
