--- app/models/aged_offspring.rb
@@ -1,3 +1,0 @@
-class AgedOffspring < Offspring
-  validates :age, presence: true, numericality: { only_integer: true, greater_than_or_equal_to: 1 }
-end
