--- app/models/graded_offspring.rb
@@ -1,4 +1,0 @@
-class GradedOffspring < Offspring
-  enum grade: %i(primary_first primary_second primary_third others)
-  validates :grade, presence: true
-end
