+++ app/models/offspring.rb
@@ -3,4 +3,6 @@ class Offspring < ApplicationRecord
   validates :first_name, presence: true, length: { within: 2..60 }
   validates :last_name, presence: true, length: { within: 2..60 }
   has_one :assignment, dependent: :destroy
+  enum grade: %i(primary_first primary_second primary_third others)
+  validates :grade, presence: true
 end
