+++ app/views/admin/offsprings/_offspring.html.haml
@@ -2,6 +2,7 @@
   %td= link_to offspring.id, admin_offspring_path(offspring)
   %td= link_to offspring.first_name, admin_offspring_path(offspring)
   %td= link_to offspring.last_name, admin_offspring_path(offspring)
+  %td= I18n.t offspring.grade, scope: 'grades'
   %td= link_to full_name(offspring.user), admin_user_path(offspring.user)
   %td= l offspring.created_at
   %td= l offspring.updated_at
