--- app/views/user/graded_offsprings/_graded_offspring.html.haml
@@ -1,33 +1,0 @@
-.list-group.list-view-pf-equalized-column
-  .list-group-item.list-view-pf-stacked
-    .list-view-pf-actions
-      %button.btn.btn-default= link_to (t '.assign'), "#"
-      .dropdown.pull-right.dropdown-kebab-pf
-        %button#dropdownKebabRight3.btn.btn-link.dropdown-toggle{"aria-expanded": "true", "aria-haspopup": "true", "data-toggle": "dropdown", type: "button"}
-          %span.fa.fa-ellipsis-v
-        %ul.dropdown-menu.dropdown-menu-right{"aria-labelledby" => "dropdownKebabRight3"}
-          %li
-            %a{:href => "#"} Edit
-          %li.divider{:role => "separator"}
-          %li
-            %a{:href => "#"} Delete
-
-    .list-view-pf-main-info
-      .list-view-pf-body
-        .list-view-pf-description
-          .list-group-item-heading
-            = graded_offspring.first_name
-          .list-group-item-text
-            = graded_offspring.last_name
-        .list-view-pf-additional-info
-          .list-view-pf-additional-info-item.list-view-pf-additional-info-item
-            = graded_offspring.grade
-          .list-view-pf-additional-info-item.list-view-pf-additional-info-item
-            - if graded_offspring.assignment
-              %strong= graded_offspring.assignment.shift_id
-            - else
-              %strong= t ".no_assignment_yet"
-          .list-view-pf-additional-info-item.list-view-pf-additional-info-item
-            - if graded_offspring.assignment
-              %strong= graded_offspring.assignment.shift.room.name
-
