+++ app/views/user/offsprings/new.html.haml
@@ -12,6 +12,6 @@
         = f.text_field :last_name, class: 'form-control'
       .form-group
         = f.label :grade
-        = f.integer :grade, class: 'form-control'
+        = f.number_field :grade, class: 'form-control'
       .actions
         = f.submit (t'.create'), class: "btn btn-primary btn-lg"
