+++ db/migrate/20161109080246_delete_type_from_offspring.rb
@@ -1,0 +1,5 @@
+class DeleteTypeFromOffspring < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :offsprings, :type, :string
+  end
+end
