--- spec/factories/graded_offspring.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :gradedOffspring, class: GradedOffspring, parent: :offspring do
-    grade { GradedOffspring.grades.keys.sample }
-  end
-end
