+++ spec/models/offspring_spec.rb
@@ -4,6 +4,10 @@ RSpec.describe Offspring, type: :model do
   i18n_scope = 'activerecord.errors.models.offspring.attributes'
   let(:offspring) { FactoryGirl.build(:offspring) }
 
+  it "has a valid factory" do
+    expect(offspring).to be_valid
+  end
+
   describe "#first_name" do
     it "is invalid without first_name" do
       offspring.first_name = ""
@@ -43,6 +47,27 @@ RSpec.describe Offspring, type: :model do
       expect(offspring.errors[:last_name]).to include(I18n.t('last_name.too_long', count: 60, scope: i18n_scope))
     end
   end
+
+  describe "#grade " do
+    it "is invalid without grade" do
+      offspring.grade = nil
+      offspring.valid?
+      expect(offspring.errors[:grade]).to include(I18n.t('grade.blank', scope: i18n_scope))
+    end
+
+    it 'is in any grade' do
+      Offspring.grades.keys.each do |i|
+        offspring.grade = i
+        offspring.valid?
+        expect(offspring).to be_valid
+      end
+    end
+
+    it "is invalid with grades that don't belong to the grade list" do
+      expect { offspring.grade = 100 }.to raise_error(ArgumentError).with_message(/is not a valid grade/)
+    end
+  end
+
   describe "#user" do
     it "is invalid wihout a user" do
       offspring.user = nil
