+++ spec/requests/user/offsprings_request_spec.rb
@@ -2,42 +2,11 @@ require 'rails_helper'
 include Warden::Test::Helpers
 
 RSpec.describe User::OffspringsController, type: :request do
-  let(:type) { Rails.application.config.offspring_type }
-  let(:type_symbol) { type&.camelize(:lower).intern }
-
-  it "is properly configured" do
-    expect(type.safe_constantize).not_to be_nil
-  end
-
-  context "when config type is wrong" do
-    let(:user) { FactoryGirl.create(:user) }
-    let(:offspring) { FactoryGirl.create(type_symbol, user: user) }
-    let(:offspring2) { FactoryGirl.create(type_symbol, user: user) }
-    let(:new_offspring) { FactoryGirl.parameters(type_symbol, user: user) }
-
-    before(:all) do
-      @offspring_type = Rails.application.config.offspring_type
-    end
-
-    after(:all) do
-      Rails.application.config.offspring_type = @offspring_type
-    end
-
-    it "returns flash message when offspring type is wrong" do
-      login_as(user, scope: :user)
-      Rails.application.config.offspring_type = "CompletelyWrong"
-      get new_user_offspring_path(user)
-      expect(response).to redirect_to(root_path)
-      expect(I18n.t('user.offsprings.new.missing_offspring_type')).not_to include "translation missing:"
-      expect(flash[:alert]).to include I18n.t "user.offsprings.new.missing_offspring_type"
-    end
-  end
-
   context "when user authenticated" do
     let(:user) { FactoryGirl.create(:user) }
-    let(:offspring) { FactoryGirl.create(type_symbol, user: user) }
-    let(:offspring2) { FactoryGirl.create(type_symbol, user: user) }
-    let(:new_offspring) { FactoryGirl.parameters(type_symbol, user: user) }
+    let(:offspring) { FactoryGirl.create(:offspring, user: user) }
+    let(:offspring2) { FactoryGirl.create(:offspring, user: user) }
+    let(:new_offspring) { FactoryGirl.parameters(:offspring, user: user) }
 
     before(:each) do
       login_as(user, scope: :user)
@@ -61,18 +30,17 @@ RSpec.describe User::OffspringsController, type: :request do
     describe "GET #new" do
       it "returns new offspring form" do
         user
-        debugger
-        get new_polymorphic_path(offspring, user.to_sym)
+        get new_user_offspring_path()
         expect(response).to have_http_status(:success)
       end
     end
 
     describe "POST #create" do
       it "creates valid offspring" do
-        expect { post polymorphic_path(user, @offspring), params: { type_symbol => FactoryGirl.attributes_for(type_symbol) } }
+        expect { post user_offspring_path(user), params: { offspring: FactoryGirl.attributes_for(:offspring) } }
           .to change(Offspring, :count).by(1)
         expect("user.offsprings.create.offspring_added").not_to include "translation missing:"
-        expect(flash[:success]).to eq I18n.t("user.offsprings.create.offspring_added", [type_symbol => offspring.first_name])
+        expect(flash[:success]).to eq I18n.t("user.offsprings.create.offspring_added", [ offspring: offspring.first_name])
         expect(response).to redirect_to admin_room_path Room.last
       end
       it "does not create invalid offspring" do
